+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   * Therion run output has no line breaks after copy/paste.
   * Searching for Therion install path on windows always failing.
   * Therion run on Windows.
+  * Windows tests.
 * Infrastructure maintenance:
   * Included "Current status" section in README.md. [requested by CaverBruce]
   * Included "Choose THConfig and run Therion" and "Run Therion" buttons in both MapiaHome and TH2FileEdit pages.